pyproject.toml
[project]
name = "datastart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datastar-py>=0.6.5",
    "nitro-boost>=0.1.0",
    "sanic>=25.3.0",
]
